+++ packages/larva-js/src/video-showcase.js
@@ -1,0 +1,9 @@
+import initCollapsible from './interface/Collapsible';
+import initFlickity from './interface/Flickity';
+import initVideoShowcase from './interface/VideoShowcase';
+
+document.addEventListener( 'DOMContentLoaded', () => {
+	initCollapsible();
+	initFlickity();
+	initVideoShowcase();
+} );