+++ packages/larva-patterns/components/c-nav-link/c-nav-link.prototype.js
@@ -1,0 +1,7 @@
+module.exports = {
+	c_nav_link_classes: '',
+	c_nav_link_text: 'Navigation Item',
+	c_nav_link_screen_reader_text: 'Navigation Item',
+	c_nav_link_url: '#',
+	c_nav_link_aria_current_attr: 'page'
+}