+++ packages/larva/lib/server.js
@@ -3,7 +3,6 @@ const express = require('express');
 const marked = require( 'marked' );
 const fs = require( 'fs' );
 const globby = require( 'globby' );
-const chalk = require( 'chalk' );
 
 const {
 	TwingEnvironment,
@@ -11,15 +10,17 @@ const {
 	TwingFilter
 } = require('twing');
 
-const getAppConfiguration = require( './utils/getAppConfiguration' );
+const getAppConfiguration = require('./utils/getAppConfiguration' );
 const getPatternPathsToLoad = require( './utils/getPatternPathsToLoad' );
 const getPatternData = require( './utils/getPatternData' );
 const getAllPatternsObj = require( './utils/getAllPatternsObj' );
 
 const app = express();
 
-const appConfiguration = getAppConfiguration( 'patterns' );
-const twigPaths = getPatternPathsToLoad( appConfiguration );
+const patternConfig = getAppConfiguration( 'patterns' );
+const brandConfig = getAppConfiguration( 'brand' );
+const assetsConfig = getAppConfiguration( 'assets' );
+const twigPaths = getPatternPathsToLoad( patternConfig );
 
 let loader = new TwingLoaderFilesystem( twigPaths );
 
@@ -34,12 +35,29 @@ const markdownFilter = new TwingFilter( 'markdown', ( string ) => {
 	}
 });
 
+const kebabify = ( name ) => {
+	let kebabCase = [];
+
+	// TODO: find a more concise way of turning a string into an iterable
+	for (let i = 0; i < name.length; i++) {
+		let letter = name[i];
+		kebabCase[i] = letter;
+	}
+
+	return kebabCase.reduce( ( a, b ) => {
+		if ( '_' === b ) {
+			return a.toLowerCase() + '-';
+		}
+		return a.toLowerCase() + b.toLowerCase();
+	});
+};
+
 // TODO: Could be an array/iterator if the namespace can be extracted from the key, the larva.config API could
 // change to `patterns: { larva: /larva/path/here/, project: /project/path/here }`
-loader.addPath( appConfiguration.larvaPatternsDir, 'larva' );
+loader.addPath( patternConfig.larvaPatternsDir, 'larva' );
 
-if( appConfiguration.projectPatternsDir ) {
-	loader.addPath( appConfiguration.projectPatternsDir, 'project' );
+if( fs.existsSync( patternConfig.projectPatternsDir ) ) {
+	loader.addPath( patternConfig.projectPatternsDir, 'project' );
 }
 
 let twing = new TwingEnvironment( loader, { debug: true } );
@@ -51,27 +69,37 @@ let patterns = {
 	project: getAllPatternsObj( appConfiguration.projectPatternsDir )
 };
 
-app.use( '/packages/' , express.static( path.join( appConfiguration.larvaPatternsDir, '../' ) ) );
-
 // NOTE: When the static site builder script is merged, this manual pattern
 // collection for the nav will come from an object based on the directory structure
 
-if( appConfiguration.larvaPatternsDir ) {
-	app.use( '/assets' , express.static( path.join( __dirname, '../' ) ) );
+if( patternConfig.larvaPatternsDir ) {
+	patterns.larva.modules = getSubDirectoryNames( path.join( patternConfig.larvaPatternsDir + '/modules' ) );
+	patterns.larva.objects = getSubDirectoryNames( path.join( patternConfig.larvaPatternsDir + '/objects' ) );
+	patterns.larva.components = getSubDirectoryNames( path.join( patternConfig.larvaPatternsDir + '/components' ) );
+	patterns.larva.tests = getSubDirectoryNames( path.join( patternConfig.larvaPatternsDir + '/__tests__' ) );
 }
 
-if( appConfiguration.projectPatternsDir ) {
-	app.use( '/assets' , express.static( path.join( appConfiguration.projectPatternsDir, '../../' ) ) );
+if( fs.existsSync( patternConfig.projectPatternsDir ) ) {
+	patterns.project.modules = getSubDirectoryNames( path.join( patternConfig.projectPatternsDir + '/modules' ) );
+	patterns.project.objects = getSubDirectoryNames( path.join( patternConfig.projectPatternsDir + '/objects' ) );
+	patterns.project.components = getSubDirectoryNames( path.join( patternConfig.projectPatternsDir + '/components' ) );
+	patterns.project.oneOffs = getSubDirectoryNames( path.join( patternConfig.projectPatternsDir + '/one-offs' ) );
+	patterns.project.tests = getSubDirectoryNames( path.join( patternConfig.projectPatternsDir + '/__tests__' ) );
+}
+
+// Use the project version if it exists, else use the larva version
+if ( fs.existsSync( assetsConfig.path ) ) {
+	app.use( '/assets' , express.static( assetsConfig.path ) );
 }
 
 app.get( '/', function (req, res) {
 	req.params[ 'source' ] = 'larva';
 	req.params[ 'pattern_nav' ] = patterns;
-	req.params[ 'name' ] = 'Welcome';
+	req.params[ 'brand' ] = req.query.tokens ? req.query.tokens : brandConfig;
 	twing.render( 'index.html', req.params ).then( output => res.end( output ) );
 });
 
-app.get( '/:source/css', function (req, res) {
+app.get( '/:source?/css', function (req, res) {
 
 	/**
 	 * Generate Larva CSS Docs
@@ -89,22 +117,25 @@ app.get( '/:source/css', function (req, res) {
 	 */
 
 	const postcss = require( 'postcss' );
-	const cssPath = path.join( process.cwd(), './node_modules/@penskemediacorp/larva-css/build/' );
+	const cssPath = path.join( process.cwd(), './node_modules/@penskemediacorp/larva/build/css/larva.css' );
 	const sassPath = path.join( process.cwd(), './node_modules/@penskemediacorp/larva-css/src/' );
 
 	const hasTokens = [
-		'u-font-family',
-		'u-color',
 		'u-background-color',
 		'u-border-color',
+		'u-color',
+		'u-font-family',
+		'u-letter-spacing',
 		'u-margin',
 		'u-padding',
 	];
 
 	const baseNames = ( () => {
 
-		let files = globby.sync( sassPath + '/**/*.scss', {
-			expandDirectories: true
+		let files = globby.sync( sassPath, {
+			expandDirectories: {
+				files: ['*.scss']
+			}
 		} );
 
 		return files.map( ( item ) => {
@@ -114,20 +145,7 @@ app.get( '/:source/css', function (req, res) {
 		} );
 	} )();
 
-	const cssFiles = globby.sync( cssPath + '/**/*.css', {
-		expandDirectories: true
-	});
-
-	const cssString = ( () => {
-		let string;
-
-		// Could use a different JS array helper here but idk
-		cssFiles.forEach( file => {
-			string += fs.readFileSync( file );
-		} );
-
-		return string;
-	 } ) ();
+	const cssString = fs.readFileSync( cssPath );
 
 	const cssRoot = postcss.parse( cssString )
 
@@ -175,17 +193,20 @@ app.get( '/:source/:type/:name/:variant?', function (req, res) {
 	let patternsPath;
 
 	if ( 'larva' === req.params.source ) {
-		patternsPath = appConfiguration.larvaPatternsDir;
+		patternsPath = patternConfig.larvaPatternsDir;
 	} else if ( 'project' === req.params.source ) {
-		patternsPath = appConfiguration.projectPatternsDir;
-	} else {
-		console.error( chalk.red.bold( 'Error loading the pattern route. \nCheck the structure of the URL. It should be: \nhttp://localhost:3000/{larva|project}/{components|objects|modules|one-offs}/{optional variant}.' ) );
+		patternsPath = patternConfig.projectPatternsDir;
+	} else if ( 'assets' === req.params.source ) {
+		return res.end();
 	}
+
 
 	// Support query parameters for conditionally loading stylesheets and scripts
 	req.params[ 'query' ] = req.query;
-	req.params[ 'data' ] = getPatternData( patternsPath, req.params );
 	req.params[ 'pattern_nav' ] = patterns;
+	req.params[ 'data' ] = undefined !== patternsPath ? getPatternData( patternsPath, req.params ) : null;
+	req.params[ 'brand' ] = req.query.tokens ? req.query.tokens : brandConfig;
+	;
 
 	if ( 'algorithms' !== req.params.type ) {
 		req.params[ 'json_pretty' ] = JSON.stringify( req.params[ 'data' ], null, '\t' );
@@ -200,6 +221,76 @@ app.get( '/:source/:type/:name/:variant?', function (req, res) {
 
 		res.end( errorMessage );
 	} );
+});
+
+app.get( '/:source?/style-guide', function (req, res ) {
+
+	const brand = req.query.tokens ? req.query.tokens : brandConfig;
+	const tokensPath = path.join( assetsConfig.path, 'build/tokens' );
+
+	const fontData = (() => {
+		try {
+			return require( path.join( tokensPath, `${brand}.typography.json` ) );
+		} catch (e) {
+			return null;
+		}
+	})();
+
+	const tokensData = (() => {
+		try {
+			return require( path.join( tokensPath, `${brand}.json` ) );
+		} catch ( e ) {
+			return null;
+		}
+	})();
+
+	const fontStyles = ( () => {
+		if ( ! fontData ) return;
+
+		return Object.keys( fontData ).map( variant => {
+			const key = kebabify( variant );
+			return {
+				name: `${key}`,
+				sizes: fontData[variant]
+			};
+		});
+	} )();
+
+	const colorsByProperty = ( () => {
+		if ( ! tokensData ) return;
+
+		const colorTokens = Object.keys( tokensData ).filter( item => item.includes( 'COLOR' ) );
+		const colorNames = colorTokens.map( token => kebabify( token ) );
+
+		return colorNames.reduce( (acc, curr) => {
+			Object.keys( acc ).forEach( key => {
+				if ( curr.startsWith( key ) ) {
+					acc[key].push( 'lrv-u-' + curr );
+				}
+			});
+
+			return acc;
+		}, {
+			color: [],
+			'background-color': [],
+			'border-color': []
+		});
+	})();
+
+	req.params[ 'name' ] = `Style Guide`;
+	req.params[ 'font_styles' ] = fontStyles;
+	req.params[ 'colors' ] = colorsByProperty;
+	req.params[ 'brand' ] = req.query.tokens ? req.query.tokens : brandConfig;
+	req.params[ 'pattern_nav' ] = patterns;
+
+	twing.render( 'style-guide/index.html', req.params ).then( output => res.end( output ) ).catch( e => {
+		const errorMessage = `Cannot render template! \n\n${e}`;
+
+		console.log( e );
+
+		res.end( errorMessage );
+	} );;
+
 });
 
 module.exports = app;