+++ packages/larva/lib/utils/getPatternData.js
@@ -2,16 +2,16 @@ const chalk = require( 'chalk' );
 const path = require( 'path' );
 /**
  * Get Pattern Data
- * 
- * @param {string} patternsPath 
+ *
+ * @param {string} patternsPath
  * @param {object} params Object containing name of pattern
  */
 function getPatternData( patternsPath, params ) {
-	
+
 	const variant = params.variant || 'prototype';
 	const patternPath = path.join( patternsPath, params.type, params.name, params.name + '.' + variant + '.js' );
 	const jsonPath = path.join( patternsPath, params.type, params.name, params.name + '.json' )
-	
+
 	try {
 		return require( patternPath );
 	} catch( e ) {