+++ packages/larva/scripts/config/gulpfile.js
@@ -1,27 +1,36 @@
 const gulp = require( 'gulp' );
-const postcss = require( 'gulp-postcss' );
-const through2 = require( 'through2' );
+const gulpClean = require( 'gulp-clean' );
+const gulpIf = require( 'gulp-if' );
+const gulpIgnore = require( 'gulp-ignore' );
+const gulpPostCss = require( 'gulp-postcss' );
+const gulpRename = require( 'gulp-rename' );
+const gulpStylelint = require( 'gulp-stylelint' );
+
 const cssnano = require( 'cssnano' );
+const Fibers = require( 'fibers' );
+const globImporter = require( 'node-sass-glob-importer' );
+const { mkdirp } = require( 'fs-extra' );
 const path = require( 'path' );
+const postCss = require( 'postcss' );
 const sass = require( 'gulp-sass' );
 
-const gulpClean = require( 'gulp-clean' );
-const gulpStylelint = require( 'gulp-stylelint' );
-const globImporter = require( 'node-sass-glob-importer' );
-
 const stylelintConfig = require( './stylelint.config' );
-
 
 /**************
 Config
 **************/
 
+const cssDest = './build/css/';
+
+sass.compiler = require( 'sass' );
+
 const sassOpts = {
+	fiber: Fibers,
 	includePaths: [
 		path.resolve( './node_modules' ),
-		path.resolve( './src/scss' )
+		path.resolve( './src/scss' ),
 	],
-	importer: globImporter()
+	importer: globImporter(),
 };
 
 const stylelintOpts = {
@@ -30,69 +39,145 @@ const stylelintOpts = {
 	reporters: [
 		{
 			formatter: 'string',
-			console: true
-		} ]
+			console: true,
+		},
+	],
 };
-
-const cssDest = './build/css/';
-
-
 
 /**************
 Functions
 **************/
 
-const stylelint = ( file ) => {
-	gulp.src( file ).pipe( gulpStylelint( stylelintOpts ) );
-};
-
-// We need an empty stream for the else condition when 
-// minifying CSS below. Reference: https://stackoverflow.com/a/30000562
-const emptyStream = () => {
-	var pass = through2.obj();
-
-	process.nextTick( pass.end.bind(pass) );
-	return pass;
-}
-
 /**
- * Build CSS
- * 
+ * Build CSS.
+ *
  * Used for both prod and dev commands.
- * 
+ *
  * Consider updating the name 'minify' to 'post' if/when
  * more Post CSS functionality is added.
- * 
- * @param {function} done 
- * @param {boolean} minify Run post CSS and minify output.
+ *
+ * @param {Function} done                      Post-build callback.
+ * @param {boolean}  minify                    Run post CSS and minify output.
+ * @param {boolean}  generateImportantVariants Whether to build variations with
+ *                                             `!important` added to all rules.
  */
-const buildScss = ( done, minify = false ) => {
+const buildScss = (
+	done,
+	minify = false,
+	generateImportantVariants = false
+) => {
+	// gulp-if and cssnano are incompatible, which gulp-if warns to expect.
+	const postCssPlugins = [];
 
-	// This is redundant, but was having issue with conditionally
-	// minifying within the same stream with 
-	// .pipe( minify ? postcss( [ cssnano() ] ) : emptyStream() )
-	// In interest of time, use separate streams for now.
-
-	if ( true === minify ) {
-		gulp.src( './entries/*.scss' )
-			.pipe( gulpStylelint( stylelintOpts ) )
-			.pipe( sass( sassOpts ).on( 'error', sass.logError ) )
-			.pipe( postcss( [ cssnano() ] ) )
-			.pipe( gulp.dest( cssDest ) );
+	if ( minify ) {
+		postCssPlugins.push( cssnano() );
 	} else {
-		gulp.src( './entries/*.scss' )
-			.pipe( gulpStylelint( stylelintOpts ) )
-			.pipe( sass( sassOpts ).on( 'error', sass.logError ) )
-			.pipe( gulp.dest( cssDest ) );
+		postCssPlugins.push( postCssNoop );
 	}
+
+	gulp.src( './entries/*.scss' )
+		.pipe( gulpStylelint( stylelintOpts ) )
+		.pipe( sass( sassOpts ).on( 'error', sass.logError ) )
+		.pipe( gulpPostCss( postCssPlugins ) )
+		.pipe( gulp.dest( cssDest ) )
+		.pipe( gulpIgnore.exclude( 'larva-ui.css' ) )
+		.pipe(
+			gulpIf(
+				generateImportantVariants,
+				gulpPostCss( [ declareImportanceForAll ] )
+			)
+		)
+		.pipe(
+			gulpIf(
+				generateImportantVariants,
+				gulpRename( { suffix: '-important' } )
+			)
+		)
+		.pipe( gulpIf( generateImportantVariants, gulp.dest( cssDest ) ) );
 
 	done();
 };
 
+/**
+ * Prepare destination directory.
+ *
+ * @param {Function} done Function called upon completion.
+ */
 const clean = ( done ) => {
-	gulp.src( cssDest , { read: false } )
-		.pipe( gulpClean() );
+	gulp.src( cssDest, { read: false } ).pipe( gulpClean() );
+
+	mkdirp( cssDest );
 	done();
+};
+
+/**
+ * Build a task from dynamic input.
+ *
+ * @param {Function} done                      Function called upon completion.
+ * @param {boolean}  generateImportantVariants Whether to build variations with
+ *                                             `!important` added to all rules.
+ */
+const composeTask = ( done, generateImportantVariants = false ) => {
+	clean( () => {
+		stylelint( './src/**/*.scss' );
+		buildScss( done, true, generateImportantVariants );
+	} );
+};
+
+/**
+ * Add `!important` declarations to all rules, for use with older themes that
+ * have high specificity.
+ *
+ * Adapted from the `postcss-importantly` npm package, updated to work with
+ * latest version of PostCSS.
+ *
+ * @param {Object} style  PostCSS Root object for current CSS.
+ * @param {Object} result PostCSS Result object containing transformed CSS.
+ * @return {Result} PostCSS Result object.
+ */
+const declareImportanceForAll = ( style, result ) => {
+	const root = postCss.parse( style );
+
+	root.walkRules( ( rule ) => {
+		if ( 'atrule' === rule.parent.type && 'media' !== rule.parent.name ) {
+			return;
+		}
+
+		return rule.each( ( decl ) => {
+			if ( ! decl.value || decl.important ) {
+				return;
+			}
+
+			decl.value += '!important';
+
+			return decl;
+		} );
+	} );
+
+	result.root = root;
+
+	return result;
+};
+
+/**
+ * PostCSS plugin that does nothing, for conditionally minifying CSS.
+ *
+ * @param {Object} style  PostCSS Root object for current CSS.
+ * @param {Object} result PostCSS Result object containing transformed CSS.
+ * @return {Result} PostCSS Result object.
+ */
+const postCssNoop = ( style, result ) => {
+	result.root = postCss.parse( style );
+	return result;
+};
+
+/**
+ * Run stylelint.
+ *
+ * @param {string} file Filename or glob to process.
+ */
+const stylelint = ( file ) => {
+	gulp.src( file ).pipe( gulpStylelint( stylelintOpts ) );
 };
 
 /**************
@@ -100,21 +185,25 @@ Tasks
 ***************/
 
 // Watch the changed file, compile and lint when changed.
-exports['dev-scss'] = () => {
-	gulp.watch( [ './src/**/*.scss', './entries/*.scss' ], buildScss ).on( 'change', function( file ) {
-		stylelint( file );
-	} );
+exports[ 'dev-scss' ] = () => {
+	gulp.watch( [ './src/**/*.scss', './entries/*.scss' ], buildScss ).on(
+		'change',
+		function ( file ) {
+			stylelint( file );
+		}
+	);
 };
 
 // Quickly build SCSS.
-exports['build-scss'] = ( done ) => {
+exports[ 'build-scss' ] = ( done ) => {
 	buildScss( done );
 };
 
 // Run PostCSS on CSS.
-exports['prod-scss'] = ( done ) => {
-	clean( () => {
-		stylelint( './src/**/*.scss' );
-		buildScss( done, true );
-	});
+exports[ 'prod-scss' ] = ( done ) => {
+	composeTask( done );
+};
+
+exports[ 'prod-scss-with-important-variants' ] = ( done ) => {
+	composeTask( done, true );
 };