+++ packages/site/src/data.js
@@ -1,0 +1,17 @@
+export const brands = [
+	"artinamerica",
+	"artnews",
+	"deadline",
+	"default",
+	"dirt",
+	"indiewire",
+	"rollingstone",
+	"sheknows",
+	"sportico",
+	"spy",
+	"thr",
+	"variety-vip",
+	"variety",
+	"vibe",
+	"wwd",
+];