+++ packages/twig-to-php-parser/__tests__/fixtures/template-parts-expected/patterns/modules/breadcrumbs.php
@@ -1,4 +1,4 @@
 <?php
 // This is a generated file. Refer to the relevant Twig file for adjusting this markup.
 ?>
-<?php \PMC::render_template( CHILD_THEME_PATH . '/template-parts/patterns/objects/o-nav.php', $o_nav, true ); ?>
+<?php \PMC\Larva\Pattern::get_instance()->render_pattern_template( 'objects/o-nav', $o_nav, true ); ?>