+++ packages/twig-to-php-parser/__tests__/fixtures/template-parts-expected/patterns/modules/trending.php
@@ -3,9 +3,10 @@
 ?>
 <section class="u-max-width-300">
 	<header>
-		<?php \PMC::render_template( CHILD_THEME_PATH . '/template-parts/patterns/components/c-heading.php', $c_heading, true ); ?>
+		<?php \PMC\Larva\Pattern::get_instance()->render_pattern_template( 'components/c-heading', $c_heading, true ); ?>
 	</header>
 	<div class="pmc-u-margin-b-0 u-border-bottom-2">
-		<?php \PMC::render_template( CHILD_THEME_PATH . '/template-parts/patterns/objects/o-story-list.php', $o_story_list, true ); ?>
+		<?php \PMC\Larva\Pattern::get_instance()->render_pattern_template( 'objects/o-story-list', $o_story_list, true ); ?>
 	</div>
+	<?php do_action( 'trending_ads_wp_action' ); ?>
 </section>